+++ client/src/components/form-input/index.ts
@@ -1,0 +1,2 @@
+import FormInput from './FormInput'
+export default FormInput
