--- client/src/scenes/login/loginValidationSchema.ts
@@ -1,8 +1,0 @@
-import { object, string } from 'yup'
-
-const loginValidationSchema = object({
-  username: string().required('Username is required'),
-  password: string().required('Password is required'),
-})
-
-export default loginValidationSchema
