+++ client/src/schema/validationSchema.ts
@@ -1,0 +1,6 @@
+import { object, string } from 'yup'
+
+export const loginValidationSchema = object({
+  username: string().label('Username').required(),
+  password: string().label('Password').required(),
+})
